requirements.txt
Django==4.0.1
django-admin-rangefilter==0.6.3
django-allauth==0.47.0
django-autocomplete-light==3.5.1
django-ckeditor==6.2.0
django-filter==21.1
django-guardian==2.2.0
django-taggit @ git+https://github.com/jazzband/django-taggit.git@d0833966d357ed1e1b9b1f40c39c1ed99affbf9b
django-taggit-serializer==0.1.7
django-timezone-field==3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
numpy==1.21.1
pandas==1.0.1
pep8==1.7.1
Pillow==7.0.0
reportlab==3.5.68
requests==2.22.0
requests-oauthlib==1.3.0
